+++ packages/reactabular-tree/src/search.js
@@ -1,19 +1,32 @@
-import { compose } from 'redux';
 import { multipleColumns } from 'reactabular-search';
-import pack from './pack';
-import unpack from './unpack';
+import getParents from './get-parents';
 
 function searchTree({
   columns,
-  query
-}) {
-  return compose(
-    unpack(),
-    multipleColumns({
+  query,
+  parentField = 'parent'
+} = {}) {
+  // Track fetched parents to get them into the results only once
+  const fetchedParents = {};
+
+  return rows => [].concat(
+    ...multipleColumns({
       columns,
       query
-    }),
-    pack()
+    })(rows).map((row) => {
+      const rowParent = row[parentField];
+
+      if (fetchedParents[rowParent]) {
+        return row;
+      }
+
+      fetchedParents[rowParent] = true;
+
+      return getParents({
+        index: row._index,
+        parentField
+      })(rows).concat(row);
+    }).filter(a => a)
   );
 }